+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.8
 ### Added
 - Other Canadian holidays: Victoria Day, Thanksgiving, Rememberance Day.
+- Limited UK holiday support.
 ### Changed
 - Fix name of Labour Day in Canada.
 - Improve accuracy of Labor Day (`None` before 1984).