+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## 0.1.8
+## [0.1.8] - 2020-11-28
 ### Added
 - Other Canadian holidays: Victoria Day, Thanksgiving, Rememberance Day.
 - Limited UK holiday support.
@@ -70,6 +70,7 @@
 		- `next` / `last` / etc. `<UNIT>` / `<DAY OF WEEK>` / `<MONTH>` / etc.
 		- `<TIME> <TIME OF DAY AM / PM>`
 
+[0.1.8]: https://github.com/AgalmicVentures/HumanTime/compare/0.1.7...0.1.8
 [0.1.7]: https://github.com/AgalmicVentures/HumanTime/compare/0.1.6...0.1.7
 [0.1.6]: https://github.com/AgalmicVentures/HumanTime/compare/0.1.5...0.1.6
 [0.1.5]: https://github.com/AgalmicVentures/HumanTime/compare/0.1.4...0.1.5